m1

+++ iCaRL_net.py
@@ -30,7 +30,7 @@ def to_onehot(targets, batch_size, num_classes):
         temp[t] = 1
         one_hots.append(temp)
     print(one_hots)
-    return one_hots
+    return torch.FloatTensor(one_hots)
 
 
 class iCaRL(nn.Module):
